+++ backend/src/main/java/com/example/backend/model/member/comment/MemberCommentDTO.java
@@ -20,7 +20,7 @@ public class MemberCommentDTO {
 
     private Timestamp updatedAt;
 
-    private int rating;
+    private double rating;
 
     public int getId() {
         return id;
@@ -86,11 +86,11 @@ public class MemberCommentDTO {
         this.updatedAt = updatedAt;
     }
 
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(double rating) {
         this.rating = rating;
     }
 }
